refactoring: Introduces a new package: .preproc
