+++ src/main/java/org/bricolages/streaming/Application.java
@@ -1,4 +1,5 @@
 package org.bricolages.streaming;
+import org.bricolages.streaming.preproc.*;
 import org.bricolages.streaming.event.*;
 import org.bricolages.streaming.stream.*;
 import org.bricolages.streaming.stream.op.OpBuilder;
